// src/types/movie.types.d.ts
interface IMovie{
    _id:string;
    title:string;
    description:string;
    videoUrl:string;
    thumbnailUrl:string;
    genre:string;
    duration:string;
    rating:number;
    mood:string;

}

interface IRecommendedMovie{
    recommendation:string[],
    reason :string;

}


export {IMovie,IRecommendedMovie}